move subspaces contents
