+++ src/domain/communication/room/Comments/CommentInputField.tsx
@@ -10,7 +10,7 @@ import { useCombinedRefs } from '@/domain/shared/utils/useCombinedRefs';
 import { HelpOutlineOutlined } from '@mui/icons-material';
 import Gutters from '@/core/ui/grid/Gutters';
 import { useSpace } from '@/domain/space/context/useSpace';
-import { useSubSpace } from '@/domain/journey/subspace/hooks/useSubSpace';
+import { useSubSpace } from '@/domain/space/hooks/useSubSpace';
 
 export const POPPER_Z_INDEX = 1400; // Dialogs are 1300
 const MAX_USERS_LISTED = 30;
