+++ src/domain/journey/common/EntityPageLayout/index.ts
@@ -2,7 +2,7 @@ import { default as EntityPageLayoutImpl } from './EntityPageLayout';
 import createLayoutHolderWithOutlet from '@/core/ui/layout/layoutHolder/LayoutHolderWithOutlet';
 import { EntityPageLayoutProps, EntityTabsProps } from './EntityPageLayoutTypes';
 import TopLevelLayout from '@/main/ui/layout/TopLevelLayout';
-import SubspacePageLayoutImpl from '@/domain/journey/subspace/layout/SubspacePageLayout';
+import SubspacePageLayoutImpl from '@/domain/space/layout/flowLayout/SubspacePageLayout';
 
 const {
   LayoutHolder: EntityPageLayoutHolder,
