+++ src/domain/space/admin/OpportunitySettings/OpportunitySettingsView.tsx
@@ -9,7 +9,7 @@ import { useDeleteSpaceMutation, useSpacePrivilegesQuery } from '@/core/apollo/g
 import { AuthorizationPrivilege } from '@/core/apollo/generated/graphql-schema';
 import DeleteIcon from '@/domain/shared/icons/DeleteIcon';
 import EntityConfirmDeleteDialog from '@/domain/shared/components/EntityConfirmDeleteDialog';
-import { useSubSpace } from '@/domain/journey/subspace/hooks/useSubSpace';
+import { useSubSpace } from '@/domain/space/hooks/useSubSpace';
 import PageContent from '@/core/ui/content/PageContent';
 import { useSpace } from '@/domain/space/context/useSpace';
