+++ src/domain/space/admin/SpaceSettings/SpaceSettingsView.tsx
@@ -31,7 +31,7 @@ import { Box, Button, CircularProgress, useTheme } from '@mui/material';
 import PageContentBlockHeader from '@/core/ui/content/PageContentBlockHeader';
 import DeleteIcon from '../../../shared/icons/DeleteIcon';
 import EntityConfirmDeleteDialog from '../../../shared/components/EntityConfirmDeleteDialog';
-import { useSubSpace } from '@/domain/journey/subspace/hooks/useSubSpace';
+import { useSubSpace } from '@/domain/space/hooks/useSubSpace';
 import { useSpace } from '@/domain/space/context/useSpace';
 import Gutters from '@/core/ui/grid/Gutters';
 import CreateTemplateDialog from '@/domain/templates/components/Dialogs/CreateEditTemplateDialog/CreateTemplateDialog';
