+++ src/domain/space/components/InfoColumn.tsx
